Move `free_port` fixture to libs
