+++ tests/pathfinding/fixtures/api.py
@@ -1,5 +1,4 @@
 # pylint: disable=redefined-outer-name
-import socket
 from typing import Iterator
 
 import pytest
@@ -11,15 +10,6 @@ from raiden.utils.typing import Address
 from raiden_libs.constants import DEFAULT_API_HOST
 
 from ..utils import SimpleReachabilityContainer
-
-
-@pytest.fixture(scope="session")
-def free_port() -> int:
-    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    sock.bind(("localhost", 0))  # binding to port 0 will choose a free socket
-    port = sock.getsockname()[1]
-    sock.close()
-    return port
 
 
 @pytest.fixture(scope="session")
